pos-system/backend-django/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
PyMySQL==1.1.0
Pillow==10.1.0
django-filter==23.3
drf-yasg==1.21.7
celery==5.3.4
redis==5.0.1
django-redis==5.4.0
python-dotenv==1.0.0
pandas==2.1.3
openpyxl==3.1.2
reportlab==4.0.9
qrcode==7.4.2
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-import-export==3.3.1
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-countries==7.5.1
django-phonenumber-field==7.2.0
django-money==3.4.0
stripe==7.5.0
django-storages==1.14.2
boto3==1.34.0
django-allauth==0.57.0
django-rest-auth==0.9.5
django-rest-knox==4.2.0
django-timezone-field==6.0.1
django-crispy-forms==2.1
crispy-bootstrap5==2023.10